Crate greg

Re-exports

pub use parser::ParserError;

pub use linker::LinkerError;

Modules

io

handles printing out files

linker

handles linking the various .data and .text segments together

parser

handles parsing and lexing the program listing

printer

handles converting the abstract program representation to binary data

Structs

DataNode

abstract data block representation

InstructionNode

Abstract instruction representation

ProgramBinary

Compiled binary

ProgramTree

Abstract representation of the whole program

Enums

AssemblerError

Bl

Block annotation types

Instr

Instructions

Token

Token types

Constants

DATA_ADDRESS_OFFSET

static variable memory addresses start here

TEXT_ADDRESS_OFFSET

instruction memory addresses start here

Functions

assemble

main function of the library - takes in a program listing, outputs an Intel HEX binary